requirements.txt
PyQt6
PyOpenGL
PyOpenGL-accelerate

torch
transformers
scipy
sounddevice
soundfile
numpy
ollama
openai-whisper
requests
pillow
huggingface_hub
accelerate
librosa
termcolor
